+++ src/components/Post/PostFormFields.js
@@ -14,9 +14,9 @@ import {
   TextField,
 } from '@material-ui/core';
 import { withStyles } from '@material-ui/core/styles';
-import { CATEGORIES } from '../constants/categories';
-import type { FormPostType } from '../business/Types';
-import { formStyle } from '../style/styles';
+import { CATEGORIES } from '../../constants/categories';
+import type { FormPostType } from '../../business/Types';
+import { formStyle } from '../../style/styles';
 
 type Props = {
   postData: FormPostType,
